Refactor: Move home page logic to homeController file
